Add nix flake template
Problem: We want for the project to be as reproducible as possible.
Me, as a mac user by now cannot build it due to the fact that
haskell tooling is not very stable on this system. More concretely
in my case, while I was building it on my global environment I got
the following cryptic errors

```
streamin: src/Data/Conduit/Internal/Pipe.hs:(413,5)-(418,38): Non-exhaustive patterns in function go
```

And this was at runtime! I even tried to make a local fork of
conduit library and fix the error, but the problem is that it was
building on my machine and while I was trying to use it in this
project, I got even more cryptic and non-helpful error.

```
streamin: internal error: stg_v_ap_get
  (GHC version 8.8.4 for x86_64_apple_darwin)
  Please reporth this as a GHC bug: https://www.haskell.org/ghc/reportabug
```

So my patience went over and I've decided to build it via nix.

Solution: Add nix flakes template, cabal.project for cabal env and
shell for hls.